Refactor command-line tools to use shared header
